#fix some stuff

+++ src/simpledb/tx/recovery/RecoveryMgr.java
@@ -5,6 +5,8 @@ import simpledb.file.*;
 import simpledb.log.*;
 import simpledb.buffer.*;
 import simpledb.tx.Transaction;
+import simpledb.tx.concurrency.LockAbortException;
+
 import static simpledb.tx.recovery.LogRecord.*;
 
 /**
@@ -33,6 +35,8 @@ public class RecoveryMgr {
     * Write a commit record to the log, and flushes it to disk.
     */
    public void commit() {
+      if (Thread.currentThread().isInterrupted())
+         throw new LockAbortException();
       bm.flushAll(txnum);
       int lsn = CommitRecord.writeToLog(lm, txnum);
       lm.flush(lsn);
